requirements.txt
allure-pytest
pytest==5.4.1
selenium==3.141.0